Revert memory permissions on cleanup
When cleanup option is specified, only freeing memory is not enough because in the
output of the /proc/$pid/maps you can see that heap is split into multiple memory regions.
To actually leave no trace it is important that we revert the allocated memory region to its
original permissions(rw-p for the heap) by calling mprotect once more.